sfjk

+++ volunteersdilemma1/dilemma1/models.py
@@ -186,7 +186,7 @@ class Player(BasePlayer):
         widget=widgets.RadioSelect
     )
 
-    bonus = models.IntegerField(initial=0, blank=True, null=True)
+    bonus = models.FloatField(initial=0, blank=True, null=True)
